requirements.txt
requests>=2.31.0
pandas>=2.0.0
plotly>=5.18.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0
duckdb>=0.10.0
apscheduler>=3.10.0
structlog>=24.1.0

# Launcher/Packaging dependencies
pystray>=0.19.0
Pillow>=10.0.0
pyinstaller>=6.0.0
